refactor: move code back to header.tsx
